requirements.txt
azure-functions
azure-storage
pandas
tables
geojson